+++ contracts/finance/andromeda-weighted-distribution-splitter/src/testing/tests.rs
@@ -1,3 +1,5 @@
+use andromeda_std::common::expiration::Expiry;
+use andromeda_std::common::Milliseconds;
 use andromeda_std::testing::mock_querier::{mock_dependencies_custom, MOCK_KERNEL_CONTRACT};
 use andromeda_std::{
     ado_base::InstantiateMsg as BaseInstantiateMsg, ado_contract::ADOContract,
@@ -7,9 +9,8 @@ use cosmwasm_std::QuerierWrapper;
 use cosmwasm_std::{
     attr,
     testing::{mock_env, mock_info},
-    Response, Timestamp, Uint128,
+    Response, Uint128,
 };
-use cw_utils::Expiration;
 
 use crate::{
     contract::{execute, instantiate},
@@ -120,19 +121,22 @@ fn test_execute_update_lock() {
     let env = mock_env();
 
     let current_time = env.block.time.seconds();
-    let lock_time = 100_000;
+    // 2 days in milliseconds
+    let lock_time = Milliseconds(172800000);
 
     let owner = "creator";
 
     // Start off with an expiration that's behind current time (expired)
     let splitter = Splitter {
         recipients: vec![],
-        lock: Expiration::AtTime(Timestamp::from_seconds(current_time - 1)),
+        lock: Milliseconds(current_time - 1),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
 
-    let msg = ExecuteMsg::UpdateLock { lock_time };
+    let msg = ExecuteMsg::UpdateLock {
+        lock_time: Expiry::FromNow(lock_time),
+    };
     ADOContract::default()
         .instantiate(
             &mut deps.storage,
@@ -153,7 +157,9 @@ fn test_execute_update_lock() {
     let info = mock_info(owner, &[]);
     let res = execute(deps.as_mut(), env.clone(), info, msg).unwrap();
 
-    let new_lock = Expiration::AtTime(Timestamp::from_seconds(current_time + lock_time));
+    let new_lock = lock_time
+        .plus_seconds(current_time)
+        .plus_milliseconds(Milliseconds(879));
     assert_eq!(
         Response::default().add_attributes(vec![
             attr("action", "update_lock"),
@@ -173,19 +179,21 @@ fn test_execute_update_lock_too_short() {
     let env = mock_env();
 
     let current_time = env.block.time.seconds();
-    let lock_time = 1;
+    let lock_time = Milliseconds(1);
 
     let owner = "creator";
 
     // Start off with an expiration that's behind current time (expired)
     let splitter = Splitter {
         recipients: vec![],
-        lock: Expiration::AtTime(Timestamp::from_seconds(current_time - 1)),
+        lock: Milliseconds(current_time - 1),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
 
-    let msg = ExecuteMsg::UpdateLock { lock_time };
+    let msg = ExecuteMsg::UpdateLock {
+        lock_time: Expiry::FromNow(lock_time),
+    };
     ADOContract::default()
         .instantiate(
             &mut deps.storage,
@@ -214,19 +222,22 @@ fn test_execute_update_lock_too_long() {
     let env = mock_env();
 
     let current_time = env.block.time.seconds();
-    let lock_time = 1_000_000_000;
+    // 25 months
+    let lock_time = Milliseconds(65_743_650_000);
 
     let owner = "creator";
 
     // Start off with an expiration that's behind current time (expired)
     let splitter = Splitter {
         recipients: vec![],
-        lock: Expiration::AtTime(Timestamp::from_seconds(current_time - 1)),
+        lock: Milliseconds(current_time - 1),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
 
-    let msg = ExecuteMsg::UpdateLock { lock_time };
+    let msg = ExecuteMsg::UpdateLock {
+        lock_time: Expiry::FromNow(lock_time),
+    };
     let deps_mut = deps.as_mut();
     ADOContract::default()
         .instantiate(
@@ -256,19 +267,22 @@ fn test_execute_update_lock_already_locked() {
     let env = mock_env();
 
     let current_time = env.block.time.seconds();
-    let lock_time = 100_000;
+
+    let lock_time = Milliseconds(172800000);
 
     let owner = "creator";
 
     // Start off with an expiration that's ahead current time (unexpired)
     let splitter = Splitter {
         recipients: vec![],
-        lock: Expiration::AtTime(Timestamp::from_seconds(current_time + 1)),
+        lock: Milliseconds::default().plus_seconds(current_time + 10_000),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
 
-    let msg = ExecuteMsg::UpdateLock { lock_time };
+    let msg = ExecuteMsg::UpdateLock {
+        lock_time: Expiry::FromNow(lock_time),
+    };
     let deps_mut = deps.as_mut();
     ADOContract::default()
         .instantiate(
@@ -298,10 +312,10 @@ fn test_execute_update_lock_unauthorized() {
     let env = mock_env();
 
     let current_time = env.block.time.seconds();
-    let lock_time = 100_000;
+    let lock_time = Milliseconds(100_000);
 
     let owner = "creator";
-    let new_lock = Expiration::AtTime(Timestamp::from_seconds(current_time - 1));
+    let new_lock = Milliseconds(current_time - 1);
 
     let splitter = Splitter {
         recipients: vec![],
@@ -310,7 +324,9 @@ fn test_execute_update_lock_unauthorized() {
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
 
-    let msg = ExecuteMsg::UpdateLock { lock_time };
+    let msg = ExecuteMsg::UpdateLock {
+        lock_time: Expiry::FromNow(lock_time),
+    };
     let deps_mut = deps.as_mut();
     ADOContract::default()
         .instantiate(
@@ -381,7 +397,7 @@ fn test_execute_remove_recipient() {
     };
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -406,7 +422,7 @@ fn test_execute_remove_recipient() {
                 weight: Uint128::new(60),
             },
         ],
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
     assert_eq!(expected_splitter, splitter);
     assert_eq!(
@@ -473,7 +489,7 @@ fn test_execute_remove_recipient_not_on_list() {
     };
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -536,16 +552,17 @@ fn test_execute_remove_recipient_contract_locked() {
     };
     let splitter = Splitter {
         recipients: recipient.clone(),
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
 
     let _res = execute(deps.as_mut(), env.clone(), info.clone(), msg).unwrap();
 
+    let current_time = env.block.time.seconds();
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(env.block.time.seconds() + 1)),
+        lock: Milliseconds::default().plus_seconds(current_time + 10_000),
     };
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
 
@@ -658,7 +675,7 @@ fn test_update_recipient_weight() {
     };
     let splitter = Splitter {
         recipients: recipient.clone(),
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -668,7 +685,7 @@ fn test_update_recipient_weight() {
     // Works
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -699,7 +716,7 @@ fn test_update_recipient_weight() {
                 weight: Uint128::new(50),
             },
         ],
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
     assert_eq!(expected_splitter, splitter);
 }
@@ -751,7 +768,7 @@ fn test_update_recipient_weight_locked_contract() {
     let current_time = env.block.time.seconds();
     let splitter = Splitter {
         recipients: recipient.clone(),
-        lock: Expiration::AtTime(Timestamp::from_seconds(current_time - 1)),
+        lock: Milliseconds(current_time - 1),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -761,7 +778,7 @@ fn test_update_recipient_weight_locked_contract() {
     // Locked contract
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(current_time + 1)),
+        lock: Milliseconds::default().plus_seconds(current_time + 1),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -828,7 +845,7 @@ fn test_update_recipient_weight_user_not_found() {
     };
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -900,7 +917,7 @@ fn test_update_recipient_weight_invalid_weight() {
     };
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -964,7 +981,7 @@ fn test_execute_add_recipient() {
     };
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1006,7 +1023,7 @@ fn test_execute_add_recipient() {
                 weight: Uint128::new(100),
             },
         ],
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
     assert_eq!(expected_splitter, splitter);
 
@@ -1069,7 +1086,7 @@ fn test_execute_add_recipient_duplicate_recipient() {
     };
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1147,7 +1164,7 @@ fn test_execute_add_recipient_invalid_weight() {
     };
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1210,9 +1227,10 @@ fn test_execute_add_recipient_locked_contract() {
         )
         .unwrap();
     let info = mock_info(owner, &[]);
+    let current_time = env.block.time.seconds();
     let splitter = Splitter {
         recipients: recipient,
-        lock: Expiration::AtTime(Timestamp::from_seconds(env.block.time.seconds() + 1)),
+        lock: Milliseconds::default().plus_seconds(current_time + 1),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1276,7 +1294,7 @@ fn test_execute_update_recipients() {
 
     let splitter = Splitter {
         recipients: vec![],
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1347,7 +1365,7 @@ fn test_execute_update_recipients_invalid_weight() {
 
     let splitter = Splitter {
         recipients: vec![],
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1402,7 +1420,7 @@ fn test_execute_update_recipients_contract_locked() {
 
     let splitter = Splitter {
         recipients: vec![],
-        lock: Expiration::AtTime(Timestamp::from_seconds(current_time + 1)),
+        lock: Milliseconds::default().plus_seconds(current_time + 10),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1455,7 +1473,7 @@ fn test_execute_update_recipients_unauthorized() {
 
     let splitter = Splitter {
         recipients: vec![],
-        lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+        lock: Milliseconds::default(),
     };
 
     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1512,7 +1530,7 @@ fn test_execute_update_recipients_unauthorized() {
 
 //     let splitter = Splitter {
 //         recipients: recipient,
-//         lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+//         lock: Milliseconds::default(),
 //     };
 
 //     let info = mock_info(owner, &[Coin::new(10000_u128, "uluna")]);
@@ -1566,7 +1584,7 @@ fn test_execute_update_recipients_unauthorized() {
 //     let env = mock_env();
 //     let splitter = Splitter {
 //         recipients: vec![],
-//         lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+//         lock: Milliseconds::default(),
 //     };
 
 //     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1592,7 +1610,7 @@ fn test_execute_update_recipients_unauthorized() {
 //     };
 //     let splitter = Splitter {
 //         recipients: vec![user1, user2],
-//         lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+//         lock: Milliseconds::default(),
 //     };
 
 //     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1650,7 +1668,7 @@ fn test_execute_update_recipients_unauthorized() {
 //     );
 //     let splitter = Splitter {
 //         recipients: recipient.clone(),
-//         lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+//         lock: Milliseconds::default(),
 //     };
 
 //     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();
@@ -1665,7 +1683,7 @@ fn test_execute_update_recipients_unauthorized() {
 //     let info = mock_info(owner, &[]);
 //     let splitter = Splitter {
 //         recipients: recipient,
-//         lock: Expiration::AtTime(Timestamp::from_seconds(0)),
+//         lock: Milliseconds::default(),
 //     };
 
 //     SPLITTER.save(deps.as_mut().storage, &splitter).unwrap();